add tweetModel. ChatToUsers, MessageModel, fix bugs

+++ src/main/java/app/model/FollowersModel.java
@@ -1,19 +1,15 @@
 package app.model;
 
-import lombok.Builder;
 import lombok.Data;
-
 import javax.persistence.*;
 
 @Entity
 @Table(name = "folowers")
 @Data
 public class FollowersModel extends BaseEntityModel{
-
     @Column(name = "follower")
-    private Integer follower_id;
-
+    private Long follower_id;
     @Column (name = "followered")
-    private Integer followered_id;
+    private Long followered_id;
 }
